requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.5
Flask-Login==0.6.3
Flask-Limiter==3.5.0
SQLAlchemy==2.0.36
redis==5.0.1
Flask-Session==0.5.0
python-socketio==5.10.0
python-engineio==4.8.0
bcrypt==4.1.2
celery==5.3.4
pytest==7.4.3
pytest-cov==4.1.0
hypothesis==6.92.1
pytest-flask==1.3.0
pytest-flask-sqlalchemy==1.1.0
fakeredis==2.20.1
pytest-mock==3.15.1
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1
